// frontend/src/app/models/player.model.ts
export interface PlayerModel {
  Number_Licenses: number;
  Name_Club: string;
  First_Name: string;
  Surname: string;
  Patronymic: string;
  Citizenship: string;
  Update_Date: Date;
  Birthday: Date;
  Age: number;
  Cost: number;
  Salary: number;
  Game_Experience: number;
  Height: number;
  Weight: number;
  Kicking_Leg: string;
  Agent_Name: string;
  Agent_Phone: string;
}